pyproject.toml
[project]
name = "adoption-accelerator"
version = "0.1.0"
description = "Exploratory Data Analysis for the PetFinder Adoption Prediction dataset."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Pedro Markovicz" }]

dependencies = []

[project.optional-dependencies]
api = [
    "pandas>=2.2", "numpy>=1.26", "pyarrow>=15.0",
    "scikit-learn>=1.5", "lightgbm>=4.3.0", "xgboost>=2.0", "catboost>=1.2",
    "shap>=0.45", "sentence-transformers>=3.0", "langdetect>=1.0.9",
    "langgraph>=0.2", "python-dotenv>=1.0", "openai>=1.99",
    "fastapi>=0.110", "uvicorn[standard]>=0.29", "httpx>=0.27",
    "joblib>=1.4", "pydantic>=2.0", "pyyaml>=6.0", "scipy>=1.12",
]
streamlit = [
    "streamlit>=1.42", "requests>=2.31", "httpx>=0.27",
    "pydantic>=2.0", "python-dotenv>=1.0", "plotly>=5.22",
]
dev = [
    "pytest>=8.0", "psutil>=5.9",
]

[project.urls]
Dataset = "https://www.kaggle.com/competitions/petfinder-adoption-prediction/overview"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-p no:asyncio"

[tool.setuptools.package-dir]
adoption_accelerator = "src/adoption_accelerator"
agents = "agents"
app = "app"

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["adoption_accelerator*", "agents*", "app*"]